Renamed folder-panel to menu-panel
