+++ frontend/src/app/components/notebook/notebook.component.html
@@ -1,27 +1,16 @@
 <mat-drawer-container class="example-container" [hasBackdrop]="hasBackDrop">
-	<!-- Menu elements -->
-	<mat-drawer
-		#drawer
-		[mode]="menuMode"
-		(openedStart)="showOverlay()"
-		(closedStart)="hideOverlay()"
-	>
+
+  <!-- Menu elements -->
+	<mat-drawer #drawer [mode]="menuMode" (openedStart)="showOverlay()" (closedStart)="hideOverlay()">
 		<mat-card id="card">
-			<button
-				mat-icon-button
-				id="closePanelBtn"
-				(click)="drawer.toggle()"
-			>
+			<button mat-icon-button id="closePanelBtn" (click)="drawer.toggle()">
 				<mat-icon>keyboard_arrow_left</mat-icon>
 			</button>
 
-			<mat-card-header id="profileCard"
-				><!-- (click)="updateProfile()" -->
+			<mat-card-header id="profileCard"><!-- (click)="updateProfile()" -->
 				<div mat-card-avatar class="header-image"></div>
 				<mat-card-title id="username">{{ username }}</mat-card-title>
-				<mat-card-subtitle id="subTitle">{{
-					degree
-				}}</mat-card-subtitle>
+				<mat-card-subtitle id="subTitle">{{ degree }}</mat-card-subtitle>
 			</mat-card-header>
 		</mat-card>
 
@@ -35,28 +24,15 @@
 				<button mat-icon-button (click)="drawer.toggle()" fxHide.gt-sm>
 					<mat-icon>menu</mat-icon>
 				</button>
-
 				<span>Smart Student</span>
 				<span class="spacer"></span>
-
-				<!-- <mat-form-field class="search">
-          <input matInput type="search" color="light" placeholder="Search">
-          <mat-icon matSuffix>search</mat-icon>
-        </mat-form-field> -->
-
-				<!-- <button mat-raised-button (click)="logout()" class="mx-4">Logout</button> -->
 			</mat-toolbar>
 		</p>
 
 		<div class="container-fluid">
-			<div
-				class="row"
-				fxLayout="row"
-				fxLayout.xs="column"
-				fxLayoutGap="0px"
-			>
-				<div class="sidePanel" id="folderPanelContainer" fxHide.lt-md>
-					<app-folder-panel #folderPanelComponent></app-folder-panel>
+			<div class="row" fxLayout="row" fxLayout.xs="column" fxLayoutGap="0px">
+				<div class="sidePanel" id="menuPanelContainer" fxHide.lt-md>
+					<app-menu-panel #menuPanelComponent></app-menu-panel>
 				</div>
 				<div class="sidePanel" id="notePanelContainer" fxHide.lt-md>
 					<app-notes-panel #notePanelComponent></app-notes-panel>
