+++ frontend/src/app/components/notebook/notebook.component.scss
@@ -25,7 +25,7 @@
   // margin-left: -13%;
 }
 
-#folderPanelContainer{
+#menuPanelContainer{
   padding-right: 0px;
 }
 
@@ -58,10 +58,6 @@
 
 #tabs{
   margin-top: -20px;
-}
-
-.search{
-  font-size: medium;
 }
 
 #notebookTitle{
