+++ frontend/src/app/components/panels/menu-panel/menu-panel.component.html
@@ -1,0 +1,36 @@
+<mat-sidenav-container id="container" style=" height: 89vh; width: 40px; background-color: white; margin-right: 0px !important;" >
+	<mat-sidenav #sidenav mode="side" style="width: 100%">
+		<div id="panelContent">
+			<mat-card id="card">
+
+        <button mat-icon-button id="closePanelBtn" (click)="sidenav.toggle(); openedCloseToggle()" >
+					<mat-icon>keyboard_arrow_left</mat-icon>
+				</button>
+
+				<mat-card-header id="profileCard" (click)="updateProfile()">
+					<div mat-card-avatar class="header-image"></div>
+					<mat-card-title id="username">{{ username }}</mat-card-title>
+					<mat-card-subtitle id="subTitle">{{ bio }}</mat-card-subtitle>
+				</mat-card-header>
+			</mat-card>
+
+			<app-tree-view #treeViewComponent id="treeViewComponent"></app-tree-view>
+
+			<mat-form-field class="search">
+        <input matInput type="search" color="light" placeholder="Search" />
+				<mat-icon matSuffix>search</mat-icon>
+      </mat-form-field>
+
+      <button mat-raised-button (click)="logout()" class="mx-4">Logout</button>
+
+		</div>
+	</mat-sidenav>
+
+
+	<mat-sidenav-content class="content">
+		<!-- <p><button mat-button (click)="sidenav.toggle()">sidenav.toggle()</button></p> -->
+		<button mat-icon-button id="openPanelBtn" (click)="sidenav.toggle(); openedCloseToggle()">
+			<mat-icon>keyboard_arrow_right</mat-icon>
+		</button>
+	</mat-sidenav-content>
+</mat-sidenav-container>
