+++ frontend/src/app/components/panels/menu-panel/menu-panel.component.scss
@@ -102,3 +102,7 @@
 ::-webkit-scrollbar-thumb:hover {
   background: #555;
 }
+
+.search{
+  font-size: medium;
+}
